+++ src/main/java/com/researchspace/dryad/model/DryadEmbedded.java
@@ -1,4 +1,4 @@
-package model;
+package com.researchspace.dryad.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;