+++ src/main/java/com/researchspace/dryad/model/DryadSubmission.java
@@ -1,9 +1,9 @@
-package model;
+package com.researchspace.dryad.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import model.location.DryadGeoLocation;
+import com.researchspace.dryad.model.location.DryadGeoLocation;
 
 import java.util.List;