+++ src/main/java/com/researchspace/dryad/model/location/DryadGeoLocationPoint.java
@@ -1,4 +1,4 @@
-package model.location;
+package com.researchspace.dryad.model.location;
 
 import lombok.Data;