+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [0.1.6]
+- Refactor packages to include `com.researchspace.dryad` package.
+
+## [0.1.5]
+- Added JSONAlias to `DryadFunder` to allow deserialization from Crossref API
 
 ## [0.1.42]
 - Added `fieldOfScience` to `DryadSubmission`.